fixed most warnings in windows port

+++ guikit/winapi/widgets/button.cpp
@@ -10,7 +10,7 @@ auto pButton::create() -> void {
     destroy(hwndTip);
     hwnd = CreateWindow(L"BUTTON", L"",
         WS_CHILD | WS_TABSTOP,
-        0, 0, 0, 0, button.window()->p.hwnd, (HMENU)button.id, GetModuleHandle(0), 0);
+        0, 0, 0, 0, button.window()->p.hwnd, (HMENU)(unsigned long long)button.id, GetModuleHandle(0), 0);
     
     SetWindowLongPtr(hwnd, GWLP_USERDATA, (LONG_PTR)&button);
 }
