+++ guikit/winapi/widgets/combobutton.cpp
@@ -80,7 +80,7 @@ auto pComboButton::create() -> void {
         L"COMBOBOX", L"",
         WS_CHILD | WS_TABSTOP | CBS_DROPDOWNLIST | CBS_HASSTRINGS,
         0, 0, 0, 0,
-        comboButton.window()->p.hwnd, (HMENU)comboButton.id, GetModuleHandle(0), 0
+        comboButton.window()->p.hwnd, (HMENU)(unsigned long long)comboButton.id, GetModuleHandle(0), 0
     );
     SetWindowLongPtr(hwnd, GWLP_USERDATA, (LONG_PTR)&comboButton);
 }
