+++ guikit/winapi/widgets/frame.cpp
@@ -31,7 +31,7 @@ auto pFrame::create() -> void {
     destroy(hwnd);
     hwnd = CreateWindow(L"BUTTON", L"",
         WS_CHILD | BS_GROUPBOX,
-        0, 0, 0, 0, widget.window()->p.hwnd, (HMENU)widget.id, GetModuleHandle(0), 0);
+        0, 0, 0, 0, widget.window()->p.hwnd, (HMENU)(unsigned long long)widget.id, GetModuleHandle(0), 0);
 
     SetWindowLongPtr(hwnd, GWLP_USERDATA, (LONG_PTR)&widget);
 }
